requirements.txt
# Core ML / NLP (Python 3.8 compatible pins)
datasets>=2.14.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=1.13.0
transformers>=4.30.0

# OpenAI API for GPT-4o generation and rewriting
openai>=1.0.0

# Data / Science
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.21.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
